Move executor service functions to the hz_resources
